Update panel email on email troubleshooting page (#100)
The email address for system emails has changed. It is now
`panel@m.arion2000.xyz` instead of `panel@email.arion2000.xyz`. This is
because a new email server was recently set up. It is now used for
system emails in the Gameservercloud and should also be updated in the
documentation.

---------

Signed-off-by: Arion2000 (2000Arion) <github@arion2000.xyz>
Signed-off-by: dependabot[bot] <support@github.com>
Co-authored-by: dependabot[bot] <49699333+dependabot[bot]@users.noreply.github.com>